refactor(backup): split restore collaborators
